# packages/contracts-schema/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountMeResponse": {
        "description": "``GET /account/me`` 响应。",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AccountMeData"
          }
        },
        "required": [
          "data"
        ],
        "title": "AccountMeResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "统一错误响应。",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "登录请求。\n\n``client_type`` 与 ``device_id`` 可选：桌面工作台注册设备后登录时带上，\n便于把会话与设备关联（Web 登录默认 WEB）。",
        "properties": {
          "client_type": {
            "default": "WEB",
            "title": "Client Type",
            "type": "string"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "云端控制平面 /api/v1：认证、账号、设备、状态流（M2）；心跳、配置、任务与小程序事件同步（M3）；技术日志与商户管理维持 stub（随开发者端页面交付）",
    "title": "warehouse-control-plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/account/me": {
      "get": {
        "description": "返回当前账号、商户与许可证评估结果（含离线宽限期状态）。",
        "operationId": "read_me_api_v1_account_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountMeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "当前账号上下文",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "校验账号密码并签发令牌对（Access 15 分钟 / Refresh 可轮换）。",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "账号登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "撤销当前会话；重复注销保持幂等成功（已撤销的会话也返回 200）。",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "退出登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "轮换 Refresh Token 并重签 Access Token；重放即撤销该账号全部会话。",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "刷新访问令牌",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/config": {
      "get": {
        "description": "返回商户生效配置（版本 + 内容 + SHA-256 摘要与签名）。\n\n客户端必须先校验摘要与签名再应用（spec：验签失败保留旧配置并告警）；\n无已发布配置时 ``data`` 为 null，客户端走本地缓存兜底。",
        "operationId": "get_config_api_v1_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "拉取生效配置",
        "tags": [
          "config"
        ]
      }
    },
    "/api/v1/dev/sync/inject": {
      "post": {
        "description": "dev 工具：Mock 小程序事件加密为密文信封写入 sync_envelope。\n\n仅供本地联调与端到端测试使用；生产环境一律 403 拒绝\n（不区分是否携带有效凭证，避免成为生产注入入口）。",
        "operationId": "inject_sync_event_api_v1_dev_sync_inject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncInjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncInjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mock 小程序事件注入（dev 工具）",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/v1/devices": {
      "get": {
        "description": "列出当前商户的全部设备（数据范围由令牌 tenant_id 决定）。",
        "operationId": "list_devices_api_v1_devices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "summary": "设备列表",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/api/v1/events/snapshot": {
      "get": {
        "description": "轮询降级入口：SSE 不可用时客户端每 30 秒拉一次当前状态。",
        "operationId": "events_snapshot_api_v1_events_snapshot_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "summary": "状态快照（轮询降级）",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}